expert/NinaPro/semg_repro/requirements.txt
numpy
scipy
typeguard
matplotlib
h5py
tensorflow-gpu >= 2.0
seaborn
